change column type

+++ baidupan/model.go
@@ -9,8 +9,8 @@ type QrcodeData struct {
 	UserCode        string `json:"user_code"`
 	VerificationUrl string `json:"verification_url"`
 	QrcodeURL       string `json:"qrcode_url"`
-	ExpiresIn       int    `json:"expires_in"`
-	Interval        int    `json:"interval"`
+	ExpiresIn       uint64 `json:"expires_in"`
+	Interval        uint64 `json:"interval"`
 }
 
 type Response struct {
@@ -24,7 +24,7 @@ type UserInfo struct {
 	BaiduName   string `json:"baidu_name"`
 	NetdiskName string `json:"netdisk_name"`
 	Uk          uint64 `json:"uk"`
-	VipType     int    `json:"vip_type"`
+	VipType     uint64 `json:"vip_type"`
 }
 
 type FileListItem struct {
@@ -52,26 +52,26 @@ type FileMetasRequest struct {
 }
 
 type FileMetaItem struct {
-	Category    int    `json:"category"`
-	DateTaken   int    `json:"date_taken"`
+	Category    uint64 `json:"category"`
+	DateTaken   uint64 `json:"date_taken"`
 	Dlink       string `json:"dlink"` // expire 8H
 	Filename    string `json:"filename"`
 	FsID        int64  `json:"fs_id"`
-	Height      int    `json:"height"`
-	Isdir       int    `json:"isdir"`
+	Height      uint64 `json:"height"`
+	Isdir       uint64 `json:"isdir"`
 	Md5         string `json:"md5"`
-	OperID      int    `json:"oper_id"`
+	OperID      uint64 `json:"oper_id"`
 	Path        string `json:"path"`
-	ServerCtime int    `json:"server_ctime"`
-	ServerMtime int    `json:"server_mtime"`
-	Size        int    `json:"size"`
+	ServerCtime uint64 `json:"server_ctime"`
+	ServerMtime uint64 `json:"server_mtime"`
+	Size        uint64 `json:"size"`
 	Thumbs      struct {
 		Icon string `json:"icon"`
 		URL1 string `json:"url1"`
 		URL2 string `json:"url2"`
 		URL3 string `json:"url3"`
 	} `json:"thumbs"`
-	Width int `json:"width"`
+	Width uint64 `json:"width"`
 }
 
 type FileMetasResponse struct {
